Improve code maintainabbility
